+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/CustomerSheet/CustomerSheetConfiguration.swift
@@ -69,9 +69,6 @@ extension CustomerSheet {
             }
         }
 
-        /// If true, use the new inline card brand selector UI instead of a dropdown
-        @_spi(STP) public var enableCBCRedesign: Bool = false
-
         /// This is an experimental feature that may be removed at any time.
         /// If true (the default), the customer can delete all saved payment methods.
         /// If false, the customer can't delete if they only have one saved payment method remaining.